+++ src/components/Comment/CommentElement.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-const CommentElement = ( props ) => {
+const CommentElement = (props) => {
   const { comment, handleCommentDelete } = props;
   const [content, setContent] = useState(comment.content);
   const [isEdit, setIsEdit] = useState(false);
@@ -15,33 +15,36 @@ const CommentElement = ( props ) => {
   const onClickEdit = () => {
     setIsEdit(!isEdit);
     // add api call for editing comment here
-  }
+  };
 
   const onClickDelete = () => {
     handleCommentDelete(comment.id);
     // add api call for deleting comment here
-  }
+  };
 
   return (
     <div className="w-full flex justify-between gap-1 mb-2">
       <div className="w-3/4">
-        {
-          isEdit ? (
-          <input className="input mr-4" value={content} onChange={(e) => setContent(e.target.value)} />
-          ) : (
-            <p className="text-lg mr-4">{content}</p>
-          )
-        }
+        {isEdit ? (
+          <input
+            className="input mr-4"
+            value={content}
+            onChange={(e) => setContent(e.target.value)}
+          />
+        ) : (
+          <p className="text-lg mr-4">{content}</p>
+        )}
         <span className="text-base mr-1 text-gray-300">
           {year}.{month}.{day}
         </span>
       </div>
       <div className="w-1/4 flex flex-row-reverse items-center">
         {!isEdit && <button onClick={() => onClickDelete()}>Del</button>}
-        <button className="mr-3" onClick={() => onClickEdit()}>{isEdit ? "Done":"Edit"}</button>
+        <button className="mr-3" onClick={() => onClickEdit()}>
+          {isEdit ? "Done" : "Edit"}
+        </button>
       </div>
     </div>
   );
-}
-
+};
 export default CommentElement;
